matrix/bot/requirements.txt
matrix-nio[e2e]>=0.20.0
aiofiles>=22.1.0
python-dotenv>=1.0.0
